requirements.txt
flask
pymongo
python-dotenv
weasyprint
jinja2
